+++ packages/webawesome/src/components/details/details.ts
@@ -268,20 +268,10 @@ export default class WaDetails extends WebAwesomeElement {
 
           <span part="icon">
             <slot name="expand-icon">
-              <wa-icon
-                library="system"
-                variant="solid"
-                name=${isRtl ? 'chevron-left' : 'chevron-right'}
-                fixed-width
-              ></wa-icon>
+              <wa-icon library="system" variant="solid" name=${isRtl ? 'chevron-left' : 'chevron-right'}></wa-icon>
             </slot>
             <slot name="collapse-icon">
-              <wa-icon
-                library="system"
-                variant="solid"
-                name=${isRtl ? 'chevron-left' : 'chevron-right'}
-                fixed-width
-              ></wa-icon>
+              <wa-icon library="system" variant="solid" name=${isRtl ? 'chevron-left' : 'chevron-right'}></wa-icon>
             </slot>
           </span>
         </summary>