+++ packages/webawesome/docs/docs/resources/changelog.md
@@ -10,6 +10,12 @@ Components with the <wa-badge variant="warning">Experimental</wa-badge> badge sh
 
 ## Next
 
+### Bug Fixes and Improvements {data-no-outline}
+
+- 🚨 BREAKING: Updated `<wa-icon>` to use Font Awesome 7 [pr:1222]
+  - Added the `auto-width` attribute to automatically size icons, since FA7 is fixed-width by default now
+  - Improved support for duotone icons in `<wa-icon>`, including custom colors, custom opacity, and opacity swapping
+  - Removed the `fixed-width` attribute as it's now the default behavior
 - Added the Hindi translation [pr:1307]
 - Fixed incorrectly named exported tooltip parts in `<wa-slider>` [pr:1277]
 - Fixed a bug in `<wa-dropdown>` that caused menus to overflow the viewport instead of resizing [issue:1267]